+++ ofrak_patch_maker/src/ofrak_patch_maker/toolchain/abstract.py
@@ -266,13 +266,15 @@ class Toolchain(ABC):
                 my_env.update(env)
                 self._logger.info(f"With env: {my_env}")
                 proc = subprocess.run(
-                    args, stdout=subprocess.PIPE, encoding="utf-8", check=True, env=my_env
+                    args, capture_output=True, encoding="utf-8", check=True, env=my_env
                 )
             else:
-                proc = subprocess.run(args, stdout=subprocess.PIPE, encoding="utf-8", check=True)
+                proc = subprocess.run(args, capture_output=True, encoding="utf-8", check=True)
         except subprocess.CalledProcessError as e:
             cmd = " ".join(args)
-            raise ValueError(f'Command "{cmd}" returned non-zero exit status {e.returncode}')
+            raise ValueError(
+                f'Command "{cmd}" returned non-zero exit status {e.returncode}.\nstdout:\n{e.output}\nstderr:\n{e.stderr}'
+            )
 
         return proc.stdout