+++ ofrak_patch_maker/CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Install toolchains (x86_64, aarch64 platforms) from source or tarball instead of relying on the Debian package manager ([#405](https://github.com/redballoonsecurity/ofrak/pull/405), [#502](https://github.com/redballoonsecurity/ofrak/pull/502), [#541](https://github.com/redballoonsecurity/ofrak/pull/541), [#562](https://github.com/redballoonsecurity/ofrak/pull/562/files))
 - Deprecate reserving `.bss` space at the FEM linking step with `create_unsafe_bss_segment` method and `unsafe_bss_segment` argument. ([#505](https://github.com/redballoonsecurity/ofrak/pull/505))
 - Determine the correct allocation behavior of sections from ELF flags rather than section names. ([#505](https://github.com/redballoonsecurity/ofrak/pull/505))
+- The PatchMaker Toolchain now includes stdout and stderr in the raised ValueError upon toolchain failure. ([#634](https://github.com/redballoonsecurity/ofrak/pull/634))
 
 ## [4.0.2](https://github.com/redballoonsecurity/ofrak/compare/ofrak-patch-maker-v.4.0.1...ofrak-patch-maker-v.4.0.2)
 ### Fixed